+++ comics/comics/duelinganalogs.py
@@ -12,7 +12,7 @@ class ComicData(ComicDataBase):
 
 class Crawler(CrawlerBase):
     history_capable_days = 35
-    schedule = 'Mo,Th'
+    schedule = 'Mo,Tu,We,Th,Fr'
     time_zone = 'US/Pacific'
 
     def crawl(self, pub_date):