+++ comics/comics/kollektivet.py
@@ -13,7 +13,7 @@ class ComicData(ComicDataBase):
 
 class Crawler(CrawlerBase):
     history_capable_date = '2013-05-01'
-    schedule = 'Mo,Tu,We,Th,Fr,Sa,Su'
+    schedule = 'Mo,Tu,We,Th,Fr,Sa'
     time_zone = 'Europe/Oslo'
 
     def crawl(self, pub_date):