+++ comics/comics/pcweenies.py
@@ -12,6 +12,7 @@ class ComicData(ComicDataBase):
 
 class Crawler(CrawlerBase):
     history_capable_days = 14
+    schedule = 'Mo,We,Fr'
     time_zone = 'US/Eastern'
 
     def crawl(self, pub_date):